powerpc/cell: Rearrange order of pmu routines
Rearrange the order of the Cell pmu routines to have dependent routines
at the top of the pmu.c file.  This change simply changes the location
of the routines, no code changes were introduced.

This reordering is required to satisfy dependencies when these routines
are converted to static routines and called indirectly through the pmu
platform abstraction layer.

Signed-off-by: Geoff Levand <geoff@infradead.org>