Webcam stops other USB devices (ehci_hcd: HC died; cleaning up)
On Wed, 21 Dec 2011, Anton Alekseev wrote:

> Alan, I've applied your patch to 3.1.5 and it seems that some troubles fixed.
> HCI doesn't die now even with too small audio latencies (died without
> this patch).
> Will continue testing.

Michael and Anton:

Below is a revised version of the original patch (it's actually a
combination of several patches that I intend to submit, assuming they
work okay).  The patch is based on the 3.3 kernel, but with a little
hand-editing it ought to work with earlier or later versions too.  I'd
appreciate any testing you can do.

Geoff, I removed the changes we added to accomodate the PS3.  I think
the driver will work okay without them, but you may see the warning
message added in ehci_poll_PSS() more often than you want.  If
necessary we can modify that routine to avoid logging the warning on
the PS3, or we could make it a one-time message.

Alan Stern