ps3-vuart: don't use PREPARE_WORK
PREPARE_[DELAYED_]WORK() are being phased out.  They have few users
and a nasty surprise in terms of reentrancy guarantee as workqueue
considers work items to be different if they don't have the same work
function.

ps3_vuart wasn't overriding the work item with multiple work functions
but was using NULL for INIT_WORK() and then single PREPARE_WORK() to
set the work function.  We can simply invoke INIT_WORK() with the work
function and remove the PREPARE_WORK() usage.

Not tested at all.

It would probably be best to route this with other related updates
through the workqueue tree.

Signed-off-by: Tejun Heo <tj@kernel.org>
Cc: Geoff Levand <geoff@infradead.org>
Cc: linuxppc-dev@lists.ozlabs.org
Cc: cbe-oss-dev@lists.ozlabs.org