local: ps3_nfs_defconfig: Disable PPC64_BIG_ENDIAN_ELF_ABI_V2
Commit 8c5fa3b5c4df3d071dab42b04b971df370d99354 (powerpc/64: Make ELFv2 the
default for big-endian builds), merged in Linux-6.5-rc1 changes the calling ABI
in a way that is incompatable with the current code for the PS3's LV1 hypervisor
calls.

See the files arch/powerpc/platforms/ps3/hvcall.S and
arch/powerpc/include/asm/lv1call.h for details of the way the current LV1 calls
are done.

This change just adds the line '# CONFIG_PPC64_BIG_ENDIAN_ELF_ABI_V2 is not set'
to the ps3_defconfig file so that the PPC64_ELF_ABI_V1 is used.

Signed-off-by: Geoff Levand <geoff@infradead.org>